Dog breed classifier completed.
In this project, you will define a Convolutional Neural Network that performs better than the average human
when given the task: identifying dog breeds. Given an image of a dog, your algorithm will produce an
estimate of the dog’s breed. If supplied an image of a human, the code will *also* produce an estimate of
the closest-resembling dog breed. Along with exploring state-of-the-art CNN models for classification, you
will make important design decisions about the user experience for your app.